+++ lib/idx/introspect.ts
@@ -12,19 +12,12 @@
  */
 
 import { makeIdxState, validateVersionConfig } from './idxState';
-import { OktaAuthInterface } from '../types';
+import { IntrospectOptions, OktaAuthInterface } from '../types';
 import { IdxResponse, isRawIdxResponse } from './types/idx-js';
 import { getOAuthDomain } from '../oidc';
 import { IDX_API_VERSION } from '../constants';
 import { httpRequest } from '../http';
 import { isAuthApiError } from '../errors';
-
-export interface IntrospectOptions {
-  withCredentials?: boolean;
-  interactionHandle?: string;
-  stateHandle?: string;
-  version?: string;
-}
 
 export async function introspect (
   authClient: OktaAuthInterface, 
@@ -34,7 +27,7 @@ export async function introspect (
   let requestDidSucceed;
 
   // try load from storage first
-  const savedIdxResponse = authClient.transactionManager.loadIdxResponse();
+  const savedIdxResponse = authClient.transactionManager.loadIdxResponse(options);
   if (savedIdxResponse) {
     rawIdxResponse = savedIdxResponse.rawIdxResponse;
     requestDidSucceed = savedIdxResponse.requestDidSucceed;